refactor: move all block commitment into worker
